Refactor keys
